+++ src/lib/Scenes/Artwork/Components/AuctionCountDownTimer.tsx
@@ -1,56 +1,16 @@
 import { TimeRemaining } from "@artsy/palette"
 import { AuctionCountDownTimer_artwork } from "__generated__/AuctionCountDownTimer_artwork.graphql"
-import { DateTime } from "luxon"
 import moment from "moment"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
+import { AuctionState } from "./CommercialInformation"
 
 interface AuctionCountDownTimerProps {
   artwork: AuctionCountDownTimer_artwork
+  auctionState: AuctionState
 }
 
-interface AuctionCountDownTimerState {
-  auctionStatus: null | "hasStarted" | "hasEnded" | "hasNotStarted"
-}
-
-export class AuctionCountDownTimer extends React.Component<AuctionCountDownTimerProps, AuctionCountDownTimerState> {
-  state = {
-    auctionStatus: null,
-  }
-
-  interval = null
-
-  componentDidMount() {
-    const { sale } = this.props.artwork
-    if (!sale) {
-      return
-    }
-
-    this.interval = setInterval(() => {
-      if (sale.liveStartAt && DateTime.local() > DateTime.fromISO(sale.liveStartAt)) {
-        this.setState({
-          auctionStatus: "hasStarted",
-        })
-      } else if (!sale.liveStartAt && DateTime.local() > DateTime.fromISO(sale.startAt)) {
-        this.setState({
-          auctionStatus: "hasStarted",
-        })
-      } else if (DateTime.local() > DateTime.fromISO(sale.endAt)) {
-        this.setState({
-          auctionStatus: "hasEnded",
-        })
-      } else {
-        this.setState({
-          auctionStatus: "hasNotStarted",
-        })
-      }
-    }, 1000)
-  }
-
-  componentWillUnmount() {
-    clearInterval(this.interval)
-  }
-
+export class AuctionCountDownTimer extends React.Component<AuctionCountDownTimerProps> {
   formatDateTime(date: string) {
     const now = moment().tz(moment.tz.guess(true))
     const dateInMoment = moment(date).tz(moment.tz.guess(true))
@@ -90,8 +50,8 @@ export class AuctionCountDownTimer extends React.Component<AuctionCountDownTimer
   }
 
   render() {
+    const { auctionState } = this.props
     const { sale } = this.props.artwork
-    const { auctionStatus } = this.state
 
     if (!sale) {
       return null
@@ -102,7 +62,7 @@ export class AuctionCountDownTimer extends React.Component<AuctionCountDownTimer
     const liveSaleHasNotStarted = sale.liveStartAt && liveStartAtDate > todaysDate
     const timerLabel = this.timerLabelText(sale.startAt, sale.liveStartAt, sale.endAt)
     const startAtDate = liveSaleHasNotStarted ? sale.liveStartAt : sale.startAt
-    const countdownEnd = auctionStatus === "hasNotStarted" ? startAtDate : sale.endAt
+    const countdownEnd = auctionState === "isPreview" ? startAtDate : sale.endAt
 
     return (
       <TimeRemaining