+++ src/lib/Scenes/Artwork/Components/CommercialButtons/__tests__/BidButton-tests.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@artsy/palette"
+import { Button, Theme } from "@artsy/palette"
 import {
   ArtworkFromAuctionPreview,
   ArtworkFromClosedAuction,
@@ -13,17 +13,37 @@ import {
 } from "lib/__fixtures__/ArtworkBidAction"
 import { renderRelayTree } from "lib/tests/renderRelayTree"
 import { merge as _merge } from "lodash"
+import { Settings } from "luxon"
+import React from "react"
 import { graphql } from "react-relay"
+import { AuctionState } from "../../CommercialInformation"
 import { BidButtonFragmentContainer as BidButton } from "../BidButton"
 
 jest.unmock("react-relay")
 
 const merge: (...args: object[]) => any = _merge
 
+const realNow = Settings.now
+const realDefaultZone = Settings.defaultZoneName
+
 describe("BidButton", () => {
-  const getWrapper = async response => {
+  beforeAll(() => {
+    Settings.defaultZoneName = "America/New_York"
+    Settings.now = () => new Date("2019-08-15T12:00:00+00:00").valueOf()
+  })
+
+  afterAll(() => {
+    Settings.now = realNow
+    Settings.defaultZoneName = realDefaultZone
+  })
+
+  const getWrapper = async (response, auctionState: AuctionState) => {
     return await renderRelayTree({
-      Component: BidButton,
+      Component: (props: any) => (
+        <Theme>
+          <BidButton {...props} auctionState={auctionState} />
+        </Theme>
+      ),
       query: graphql`
         query BidButtonTestsQuery {
           artwork(id: "auction_artwork") {
@@ -39,7 +59,7 @@ describe("BidButton", () => {
 
   describe("for closed auction", () => {
     it("does not display anything", async () => {
-      const wrapper = await getWrapper(ArtworkFromClosedAuction)
+      const wrapper = await getWrapper(ArtworkFromClosedAuction, "hasEnded")
 
       expect(wrapper.html()).toBe(null)
     })
@@ -47,21 +67,21 @@ describe("BidButton", () => {
 
   describe("for auction preview", () => {
     it("and not registered bidder", async () => {
-      const wrapper = await getWrapper(ArtworkFromAuctionPreview)
+      const wrapper = await getWrapper(ArtworkFromAuctionPreview, "isPreview")
 
       expect(wrapper.text()).toContain("Register to bid")
     })
 
     it("with bidder registration pending approval", async () => {
       const artwork = merge({}, ArtworkFromAuctionPreview, BidderPendingApproval)
-      const wrapper = await getWrapper(artwork)
+      const wrapper = await getWrapper(artwork, "isPreview")
 
       expect(wrapper.text()).toContain("Registration pending")
     })
 
     it("with registered bidder", async () => {
       const artwork = merge({}, ArtworkFromAuctionPreview, RegisteredBidder)
-      const wrapper = await getWrapper(artwork)
+      const wrapper = await getWrapper(artwork, "isPreview")
 
       expect(wrapper.text()).toContain("Registration complete")
     })
@@ -70,35 +90,35 @@ describe("BidButton", () => {
   describe("for open auction", () => {
     it("with open registration and not registered bidder ", async () => {
       const artwork = merge({}, ArtworkFromTimedAuctionRegistrationOpen, NotRegisteredToBid)
-      const wrapper = await getWrapper(artwork)
+      const wrapper = await getWrapper(artwork, "hasStarted")
 
       expect(wrapper.text()).toContain("Bid")
     })
 
     it("with closed registration and not registered bidder ", async () => {
       const artwork = merge({}, ArtworkFromTimedAuctionRegistrationClosed, NotRegisteredToBid)
-      const wrapper = await getWrapper(artwork)
+      const wrapper = await getWrapper(artwork, "hasStarted")
 
       expect(wrapper.text()).toContain("Registration closed")
     })
 
     it("with bidder registration pending approval", async () => {
       const artwork = merge({}, ArtworkFromTimedAuctionRegistrationOpen, BidderPendingApproval)
-      const wrapper = await getWrapper(artwork)
+      const wrapper = await getWrapper(artwork, "hasStarted")
 
       expect(wrapper.text()).toContain("Registration pending")
     })
 
     it("with registered bidder", async () => {
       const artwork = merge({}, ArtworkFromTimedAuctionRegistrationOpen, RegisteredBidder)
-      const wrapper = await getWrapper(artwork)
+      const wrapper = await getWrapper(artwork, "hasStarted")
 
       expect(wrapper.text()).toContain("Bid")
     })
 
     it("with registered bidder with bids", async () => {
       const artwork = merge({}, ArtworkFromTimedAuctionRegistrationOpen, RegistedBidderWithBids)
-      const wrapper = await getWrapper(artwork)
+      const wrapper = await getWrapper(artwork, "hasStarted")
 
       expect(wrapper.text()).toContain("Increase max bid")
     })
@@ -107,14 +127,14 @@ describe("BidButton", () => {
   describe("for live auction", () => {
     it("with open registration and not registered bidder ", async () => {
       const artwork = merge({}, ArtworkFromLiveAuctionRegistrationOpen, NotRegisteredToBid)
-      const wrapper = await getWrapper(artwork)
+      const wrapper = await getWrapper(artwork, "isLive")
 
       expect(wrapper.find(Button).text()).toContain("Enter live bidding")
     })
 
     it("with closed registration and not registered bidder ", async () => {
       const artwork = merge({}, ArtworkFromLiveAuctionRegistrationClosed, NotRegisteredToBid)
-      const wrapper = await getWrapper(artwork)
+      const wrapper = await getWrapper(artwork, "isLive")
 
       expect(wrapper.text()).toContain("Registration closed")
       expect(wrapper.find(Button).text()).toContain("Watch live bidding")
@@ -122,14 +142,14 @@ describe("BidButton", () => {
 
     it("with bidder registration pending approval", async () => {
       const artwork = merge({}, ArtworkFromLiveAuctionRegistrationOpen, BidderPendingApproval)
-      const wrapper = await getWrapper(artwork)
+      const wrapper = await getWrapper(artwork, "isLive")
 
       expect(wrapper.find(Button).text()).toContain("Enter live bidding")
     })
 
     it("with registered bidder", async () => {
       const artwork = merge({}, ArtworkFromLiveAuctionRegistrationOpen, RegisteredBidder)
-      const wrapper = await getWrapper(artwork)
+      const wrapper = await getWrapper(artwork, "isLive")
 
       expect(wrapper.find(Button).text()).toContain("Enter live bidding")
     })