+++ src/lib/Scenes/Artwork/Components/__tests__/AuctionPrice-tests.tsx
@@ -17,15 +17,16 @@ import { renderRelayTree } from "lib/tests/renderRelayTree"
 import React from "react"
 import { graphql } from "react-relay"
 import { AuctionPriceFragmentContainer as AuctionPrice } from "../AuctionPrice"
+import { AuctionState } from "../CommercialInformation"
 
 jest.unmock("react-relay")
 
 describe("AuctionPrice", () => {
-  const getWrapper = async response => {
+  const getWrapper = async (response, auctionState: AuctionState) => {
     return await renderRelayTree({
       Component: (props: any) => (
         <Theme>
-          <AuctionPrice {...props} />
+          <AuctionPrice {...props} auctionState={auctionState} />
         </Theme>
       ),
       query: graphql`
@@ -43,7 +44,7 @@ describe("AuctionPrice", () => {
 
   describe("for closed auction", () => {
     it("displays Auction Closed", async () => {
-      const wrapper = await getWrapper(ClosedAuctionArtwork)
+      const wrapper = await getWrapper(ClosedAuctionArtwork, "hasEnded")
 
       expect(wrapper.text()).toContain("Bidding closed")
     })
@@ -51,7 +52,7 @@ describe("AuctionPrice", () => {
 
   describe("for live sale in progress", () => {
     it("does not display anything", async () => {
-      const wrapper = await getWrapper(LiveAuctionInProgeress)
+      const wrapper = await getWrapper(LiveAuctionInProgeress, "isLive")
 
       expect(wrapper.html()).toBe(null)
     })
@@ -59,7 +60,7 @@ describe("AuctionPrice", () => {
 
   describe("for auction preview", () => {
     it("displays proper starting bid info", async () => {
-      const wrapper = await getWrapper(AuctionPreview)
+      const wrapper = await getWrapper(AuctionPreview, "isPreview")
 
       expect(wrapper.text()).toContain("Starting bid")
       expect(wrapper.text()).toContain("CHF 4,000")
@@ -68,14 +69,14 @@ describe("AuctionPrice", () => {
 
   describe("for auction preview with no start bid set", () => {
     it("displays nothing if current bid info is unavailable", async () => {
-      const wrapper = await getWrapper(AuctionPreviewNoStartingBid)
+      const wrapper = await getWrapper(AuctionPreviewNoStartingBid, "isPreview")
       expect(wrapper.html()).toBe(null)
     })
   })
 
   describe("for open auction with no reserve and no bids", () => {
     it("displays proper starting bid info", async () => {
-      const wrapper = await getWrapper(OpenAuctionNoReserveNoBids)
+      const wrapper = await getWrapper(OpenAuctionNoReserveNoBids, "hasStarted")
 
       expect(wrapper.text()).toContain("Starting bid")
       expect(wrapper.text()).toContain("$500")
@@ -84,7 +85,7 @@ describe("AuctionPrice", () => {
 
   describe("open auction with no reserve with bids present", () => {
     it("displays proper current bid info including bid count", async () => {
-      const wrapper = await getWrapper(OpenAuctionNoReserveWithBids)
+      const wrapper = await getWrapper(OpenAuctionNoReserveWithBids, "hasStarted")
       const texts = wrapper.find(Sans).map(x => x.text())
 
       expect(texts).toContain("Current bid")
@@ -95,7 +96,7 @@ describe("AuctionPrice", () => {
 
   describe("for open auction with reserve and no bids", () => {
     it("displays proper starting bid info and resserve message", async () => {
-      const wrapper = await getWrapper(OpenAuctionReserveNoBids)
+      const wrapper = await getWrapper(OpenAuctionReserveNoBids, "hasStarted")
       const texts = wrapper.find(Sans).map(x => x.text())
 
       expect(texts).toContain("Starting bid")
@@ -106,7 +107,7 @@ describe("AuctionPrice", () => {
 
   describe("for open auction with some bids and reserve not met", () => {
     it("displays current bid message inculding reserve warning", async () => {
-      const wrapper = await getWrapper(OpenAuctionReserveNotMetWithBids)
+      const wrapper = await getWrapper(OpenAuctionReserveNotMetWithBids, "hasStarted")
       const texts = wrapper.find(Sans).map(x => x.text())
 
       expect(texts).toContain("Current bid")
@@ -117,7 +118,7 @@ describe("AuctionPrice", () => {
 
   describe("for open auction with some bids and satisfied reserve", () => {
     it("displays current bid message inculding reserve met", async () => {
-      const wrapper = await getWrapper(OpenAuctionReserveMetWithBids)
+      const wrapper = await getWrapper(OpenAuctionReserveMetWithBids, "hasStarted")
       const texts = wrapper.find(Sans).map(x => x.text())
 
       expect(texts).toContain("Current bid")
@@ -128,7 +129,7 @@ describe("AuctionPrice", () => {
 
   describe("for open auction with my bid winning", () => {
     it("displays max bid and winning indicator", async () => {
-      const wrapper = await getWrapper(OpenAuctionReserveMetWithMyWinningBid)
+      const wrapper = await getWrapper(OpenAuctionReserveMetWithMyWinningBid, "hasStarted")
       const texts = wrapper.find(Sans).map(x => x.text())
 
       expect(texts).toContain("Your max: $15,000")
@@ -138,7 +139,7 @@ describe("AuctionPrice", () => {
 
   describe("for open auction with my bid losing", () => {
     it("displays max bid and losing indicator", async () => {
-      const wrapper = await getWrapper(OpenAuctionReserveMetWithMyLosingBid)
+      const wrapper = await getWrapper(OpenAuctionReserveMetWithMyLosingBid, "hasStarted")
       const texts = wrapper.find(Sans).map(x => x.text())
 
       expect(texts).toContain("Your max: $400")
@@ -148,7 +149,7 @@ describe("AuctionPrice", () => {
 
   describe("for open auction with me increasing my max bid while winning", () => {
     it("displays max bid and winning indicator", async () => {
-      const wrapper = await getWrapper(OpenAuctionReserveNotMetIncreasingOwnBid)
+      const wrapper = await getWrapper(OpenAuctionReserveNotMetIncreasingOwnBid, "hasStarted")
       const texts = wrapper.find(Sans).map(x => x.text())
 
       expect(texts).toContain("Your max: $15,000")